test/robotframework-bdd/requirement.txt
robotframework==4.0.3
robotframework-seleniumlibrary==5.1.3
robotframework-datadriver==1.0.0
selenium==3.141.0
webdrivermanager==0.9.0
bs4==4.12.2
pytest==7.3.1
pathlib==1.0.1
